chore: apply suggestion
