+++ data/src/main/java/com/my/kizzy/data/di/AppModule.kt
@@ -18,6 +18,7 @@ import com.my.kizzy.data.remote.Base
 import com.my.kizzy.data.remote.Discord
 import com.my.kizzy.data.remote.Github
 import com.my.kizzy.data.remote.Imgur
+import com.my.kizzy.data.remote.ImgurApiService
 import com.my.kizzy.data.repository.KizzyRepositoryImpl
 import com.my.kizzy.domain.repository.KizzyRepository
 import dagger.Module
@@ -94,7 +95,7 @@ object AppModule {
     @Provides
     fun providesKizzyRepository(
         apiService: ApiService,
-        imgurApiService: ApiService.ImgurApiService
+        imgurApiService: ImgurApiService
     ): KizzyRepository {
         return KizzyRepositoryImpl(apiService, imgurApiService)
     }
