+++ data/src/main/java/com/my/kizzy/data/remote/ApiResponse.kt
@@ -7,11 +7,3 @@ data class ApiResponse(
     @SerialName("id")
     val id: String
 )
-
-@Serializable
-data class ExternalAsset(
-    @SerialName("external_asset_path")
-    val externalAssetPath: String,
-    @SerialName("url")
-    val url: String
-)
