+++ data/src/main/java/com/my/kizzy/data/rpc/KizzyRPC.kt
@@ -318,7 +318,7 @@ class KizzyRPC(
                 ),
                 afk = true,
                 since = startTimestamps,
-                status = this.status
+                status = Constants.DND
             )
         )
     }
