+++ gradle/libs.versions.toml
@@ -7,15 +7,15 @@ version-code ="9"
 
 # app level
 accompanist = "0.31.0-alpha"
-activity-compose = "1.9.3"
+activity-compose = "1.9.2"
 android-svg = "1.4"
 androidx-material = "1.12.0"
 app-compat = "1.7.0"
 blankj-utilcodex = "1.31.0"
 coil = "2.1.0"
-compose = "1.7.4"
+compose = "1.7.3"
 compose-compiler = "1.4.0"
-compose-navigation = "2.8.3"
+compose-navigation = "2.8.2"
 core-ktx = "1.13.1"
 crashx = "v6.0.19"
 glide = "1.6.1"
